Livekit/RAG/requirements.txt
# LiveKit Agents Framework
livekit-agents[openai,silero,elevenlabs]~=1.0

# RAG and Vector Database
pinecone>=5.0.0
openai>=1.0.0
tiktoken>=0.5.0

# Data Processing
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-pinecone>=0.0.3

# API Services
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
numpy>=1.24.0

# Development
pytest>=7.0.0
black>=23.0.0